refactor: move orth tag and all its transitive dependencies to morphodict
